+++ resources/views/dielo.blade.php
@@ -41,11 +41,11 @@
             </div>
             <div class="row">
                 <div class="col-md-8 text-center">
-                        @if (!empty($item->iipimg_url))
-                            <a href="{!! URL::to('dielo/' . $item->id . '/zoom') !!}" data-toggle="tooltip" data-placement="top" title="{{ utrans('general.item_zoom') }}">
-                        @endif    
+                        @if ($item->has_iip)
+                            <a href="{{ route('item.zoom', ['id' => $item->id]) }}" data-toggle="tooltip" data-placement="top" title="{{ utrans('general.item_zoom') }}">
+                        @endif
                         <img src="{!! $item->getImagePath() !!}" class="img-responsive img-dielo" alt="{!! $item->getTitleWithAuthors() !!}" itemprop="image">
-                        @if (!empty($item->iipimg_url))
+                        @if ($item->has_iip)
                             </a>
                         @endif
                         <div class="row">
@@ -59,8 +59,8 @@
                             </div>
 
                             <div class="col-md-12 text-center">
-                                @if (!empty($item->iipimg_url))
-                                   <a href="{!! URL::to('dielo/' . $item->id . '/zoom') !!}" class="btn btn-default btn-outline  sans"><i class="fa fa-search-plus"></i> {{ trans('general.item_zoom') }}</a>
+                                @if ($item->has_iip)
+                                   <a href="{{ route('item.zoom', ['id' => $item->id]) }}" class="btn btn-default btn-outline  sans"><i class="fa fa-search-plus"></i> {{ trans('general.item_zoom') }}</a>
                                 @endif
                                 @if ($item->isForReproduction())
                                     <a href="{!! URL::to('dielo/' . $item->id . '/objednat')  !!}" class="btn btn-default btn-outline  sans"><i class="fa fa-shopping-cart"></i> {{ trans('dielo.item_order') }} </a>
@@ -247,11 +247,11 @@
                         <?php $related_items = App\Item::where('related_work', '=', $item->related_work)->where('author', '=', $item->author)->orderBy('related_work_order')->get() ?>
                         @if ($related_items->count() > 1)
                         <div style="position: relative; padding: 0 10px;">
-                        <div class="artworks-preview small">
-                        @foreach ($related_items as $item)
-                            <a href="{!! $item->getUrl() !!}"><img data-lazy="{!! $item->getImagePath() !!}" class="img-responsive-width " alt="{!! $item->getTitleWithAuthors() !!} "></a>
-                        @endforeach
-                        </div>
+                            @include('components.artwork_carousel', [
+                                'slick_target' => "artworks-preview",
+                                'slick_variant' => "small",
+                                'items' => $related_items,
+                            ])
                         </div>
                         @endif
                     @endif
@@ -272,16 +272,37 @@
         </div>
     </div>
 </section>
-<section class="more-items content-section">
+
+@if ($colors_used || $similar_by_color)
+<section class="content-section">
+    <div class="container">
+        <div class="row">
+            <div class="col-xs-12">
+                <h4>{{ trans('dielo.more-items_similar-colors') }}</h4>
+                @if ($colors_used)
+                @include('components.color_list', ['colors' => $colors_used])
+                @endif
+                @if ($similar_by_color)
+                    @include('components.artwork_carousel', [
+                        'slick_target' => "artworks-preview",
+                        'items' => $similar_by_color,
+                    ])
+                @endif
+            </div>
+        </div>
+    </div>
+</section>
+@endif
+
+<section class="more-items content-section light-grey">
     <div class="container">
         <div class="row">
             <div class="col-xs-12">
                 <h4>{{ trans('dielo.more-items_related-artworks') }}</h4>
-                <div class="artworks-preview ">
-                @foreach ($more_items as $item)
-                    <a href="{!! $item->getUrl() !!}"><img data-lazy="{!! $item->getImagePath() !!}" class="img-responsive-width " alt="{!! $item->getTitleWithAuthors() !!} "></a>
-                @endforeach
-                </div>
+                @include('components.artwork_carousel', [
+                    'slick_target' => "artworks-preview",
+                    'items' => $more_items,
+                ])
             </div>
         </div>
     </div>
@@ -326,9 +347,10 @@
 
 
 @section('javascript')
-{!! Html::script('js/slick.js') !!}
 {!! Html::script('js/readmore.min.js') !!}
 {!! Html::script('js/jquery.fileDownload.js') !!}
+
+@include('components.artwork_carousel_js', ['slick_query' => '.artworks-preview'])
 
 @if (!empty($item->lat) && ($item->lat > 0)) 
     <!-- Google Maps API Key - You will need to use your own API key to use the map feature -->
@@ -380,17 +402,6 @@
                 }
             });
             return false; //this is critical to stop the click event which will trigger a normal file download!
-        });
-
-        $('.artworks-preview').slick({
-            dots: false,
-            lazyLoad: 'progressive',
-            infinite: false,
-            speed: 300,
-            slidesToShow: 1,
-            slide: 'a',
-            centerMode: false,
-            variableWidth: true,
         });
 
 // $(document).on("click", "#download", function() {